requirements.txt
streamlit==1.28.0
python-docx==1.1.0
pypdf2==3.0.1
fpdf2==2.7.4
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3